CODE STRUCTURE REFACTOR: Move 'FootO', 'Rogaine' and 'Relay' under 'Results'
