+++ src/pages/Results/pages/Layout/StageLayout.tsx
@@ -6,22 +6,22 @@ import {
   Paper,
   Typography,
 } from "@mui/material"
-import FootOStartTime from "../FootO/pages/StartTime/FootOStartTime.tsx"
-import FootOResults from "../FootO/pages/Results/FootOResults.tsx"
+import FootOStartTime from "../Results/pages/FootO/pages/StartTime/FootOStartTime.tsx"
+import FootOResults from "../Results/pages/FootO/pages/Results/FootOResults.tsx"
 import { AccessTime } from "@mui/icons-material"
 import EmojiEventsIcon from "@mui/icons-material/EmojiEvents"
 import TimerIcon from "@mui/icons-material/Timer"
 import TimelineIcon from "@mui/icons-material/Timeline"
-import RelayResults from "../Relay/pages/RelayResults/RelayResults.tsx"
-import RelaySplits from "../Relay/pages/RelaySplits/RelaySplits.tsx"
-import RogaineResults from "../Rogaine/pages/RogaineResults/RogaineResults.tsx"
-import RogainePoints from "../Rogaine/pages/RogainePoints/RogainePoints.tsx"
+import RelayResults from "../Results/pages/Relay/pages/RelayResults/RelayResults.tsx"
+import RelaySplits from "../Results/pages/Relay/pages/RelaySplits/RelaySplits.tsx"
+import RogaineResults from "../Results/pages/Rogaine/pages/RogaineResults/RogaineResults.tsx"
+import RogainePoints from "../Results/pages/Rogaine/pages/RogainePoints/RogainePoints.tsx"
 import { useFetchClasses, useRunners, useSelectedMenu } from "../../shared/hooks.ts"
 import ClassSelector from "./components/ClassSelector.tsx"
 import { useTranslation } from "react-i18next"
-import FootOSplits from "../FootO/pages/Splits/FootOSplits.tsx"
+import FootOSplits from "../Results/pages/FootO/pages/Splits/FootOSplits.tsx"
 import { RunnersContext, SelectedClassContext } from "../../shared/context.ts"
-import EventStageBanner from "./components/EventStageBanner.tsx"
+import EventStageBanner from "../Results/components/EventStageBanner.tsx"
 import RefreshIcon from "@mui/icons-material/Refresh"
 import Tooltip from "@mui/material/Tooltip"
 import { useFetchEventDetail } from "../../services/FetchHooks.ts"
