+++ src/pages/Results/pages/Results/pages/FootO/pages/StartTime/FootOStartTime.tsx
@@ -1,14 +1,14 @@
 import { Box, Stack, Typography } from "@mui/material"
 import { useContext, useEffect, useState } from "react"
-import { RunnersContext } from "../../../../shared/context.ts"
+import { RunnersContext } from "../../../../../../shared/context.ts"
 import { orderRunnersByStartTime } from "./shared/functions.ts"
 import { useTranslation } from "react-i18next"
-import ResultListContainer from "../../../../components/ResultsList/ResultListContainer.tsx"
-import ResultListItem from "../../../../components/ResultsList/ResultListItem.tsx"
-import StartTime from "../../../../components/StartTime.tsx"
-import { ProcessedRunnerModel } from "../../../../components/VirtualTicket/shared/EntityTypes.ts"
-import { parseResultStatus } from "../../../../shared/functions.ts"
-import ResultsListSkeleton from "../../../../components/ResultsList/ResultListSkeleton.tsx"
+import ResultListContainer from "../../../../../../components/ResultsList/ResultListContainer.tsx"
+import ResultListItem from "../../../../../../components/ResultsList/ResultListItem.tsx"
+import StartTime from "../../../../../../components/StartTime.tsx"
+import { ProcessedRunnerModel } from "../../../../../../components/VirtualTicket/shared/EntityTypes.ts"
+import { parseResultStatus } from "../../../../../../shared/functions.ts"
+import ResultsListSkeleton from "../../../../../../components/ResultsList/ResultListSkeleton.tsx"
 
 export default function FootOStartTime() {
   const { t } = useTranslation()
