+++ src/pages/Results/pages/Results/pages/Rogaine/components/RogaineVirtualTicket/components/RogaineVirtualTicketPointsBanner.tsx
@@ -1,10 +1,10 @@
 import React from "react"
 import Grid from "@mui/material/Grid"
 import { Typography } from "@mui/material"
-import { ProcessedRunnerResultModel } from "../../../../../components/VirtualTicket/shared/EntityTypes.ts"
-import { parseResultStatus } from "../../../../../shared/functions.ts"
-import RaceTime from "../../../../../components/RaceTime.tsx"
-import StartTime from "../../../../../components/StartTime.tsx"
+import { ProcessedRunnerResultModel } from "../../../../../../../components/VirtualTicket/shared/EntityTypes.ts"
+import { parseResultStatus } from "../../../../../../../shared/functions.ts"
+import RaceTime from "../../../../../../../components/RaceTime.tsx"
+import StartTime from "../../../../../../../components/StartTime.tsx"
 import { useTranslation } from "react-i18next"
 
 type RogaineVirtualTicketPointsBannerProps = {
