+++ src/pages/Results/pages/Results/pages/Rogaine/pages/RogainePoints/components/ControlBadge.tsx
